Avoid `caller` unless necessary
* Only call `caller` when transaction is sampled

`caller` is very slow on JRuby. I couldn't find a way around it, so this
changes it to only be called when a transaction/span is being sampled.
Previously, setting `transaction_sample_rate` didn't improve performance
for our JRuby app.

* Build stacktrace on worker thread

I didn't see a reason for stacktrace building/parsing to run when
generating a span on the main thread. This moves it be being generated
lazily in a worker.

* Keep old backtrace api

* Update changelog