Lazy load ActiveRecord adapter in SQL normalizer
When Rails is present the ActiveRecord's normalizer will be called during
the initialization phrase which will force ActiveRecord to make a connection.
While this is OK in most cases it's not the best approach to query the database
during the initialization phrase and it will break in some weird
scenarios where the database is not available during boot, like when you
rely on some configuration that holds the database information and is
dynamically loaded during boot.