Add Ruby interpreter check to handle non-MRI environments
This commit also restores the ActiveRecord::Base fallback logic to be
used in cases where the Object ID lookup is not possible.

It should be noted that the YARV interpreter will also cause the check
to succeed, not just MRI, but due to the complex nature of the
relationship between YARV and MRI (they are sort of the same thing?),
this is not an issue.