Close the read pipe at the right moment
We can't close the read pipe inside the child thread because it won't
always be executed, so in cases where some exception happens and prevent
the child thread from being executed, there would be a file descriptor
leak. We shall close the read side of the pipe whether we executed the
child thread or not.